sdfs

+++ pages/2_📋_CasMatcher.py
@@ -18,13 +18,15 @@ def zip_outputs():
     os.remove('./outputs.zip')
     with zipfile.ZipFile('./outputs.zip', 'w') as zipf:
         for filename in os.listdir('./outputs/'):
-            file_path = './outputs/' + filename
-            zipf.write(file_path, compress_type=zipfile.ZIP_DEFLATED)
+            if filename.split('.')[-1] != 'txt':
+                file_path = './outputs/' + filename
+                zipf.write(file_path, compress_type=zipfile.ZIP_DEFLATED)
 
 def clear_folder():
-    files = os.listdir('/outputs/')
+    files = os.listdir('./outputs/')
     for file in files:
-        os.remove(os.path.join('./outputs/', file))
+        if file.split('.')[-1] != 'txt':
+            os.remove(os.path.join('./outputs/', file))
 
 def get_match_data():
     df_data['Standard'] = []
@@ -32,11 +34,12 @@ def get_match_data():
     df_data['Filename'] = []
     files = os.listdir('./outputs/')
     for file in files:
-        list_name, matching_number, filename = file.split('&')
-        if matching_number != '0':
-            df_data['Standard'].append(list_name)
-            df_data['Matching Number'].append(matching_number)
-            df_data['Filename'].append(filename)
+        if file.split('.')[-1] != 'txt':
+            list_name, matching_number, filename = file.split('&')
+            if matching_number != '0':
+                df_data['Standard'].append(list_name)
+                df_data['Matching Number'].append(matching_number)
+                df_data['Filename'].append(filename)
     df = pd.DataFrame(df_data)
     df['Matching Number'] = df['Matching Number'].astype(int)
     return df
